feat: refactor setup param opt
